feat/add global Container component
