+++ apps/web/src/components/inc/Container.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="w-full max-w-7xl px-3">
+        <slot></slot>
+    </div>
+</template>
